+++ main_test.go
@@ -1,0 +1,27 @@
+package main
+
+import (
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"net/http"
+	"net/http/httptest"
+	"strings"
+	"testing"
+)
+
+func TestMainHandler(t *testing.T) {
+	totalCount := 4
+	req := httptest.NewRequest("GET", "/cafe?count=10&city=moscow", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	require.NotEqual(t, responseRecorder.Code, http.StatusBadRequest)
+	assert.NotEmpty(t, responseRecorder.Body)
+
+	body := responseRecorder.Body.String()
+	list := strings.Split(body, ",")
+
+	assert.Len(t, list, totalCount)
+}
